Speech-To-Text/DeepSpeech/requirements.txt
deepspeech-gpu
librosa